PSRS/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-background-task==0.1.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-framework==0.1.0
djangorestframework==3.12.2
pkg-resources==0.0.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
